Consolidate OIDC storage account and cluster MSI key vault into one ARM template + `make generate`
