+++ pkg/deploy/generator/resources_miwi_dev.go
@@ -6,6 +6,7 @@ package generator
 import (
 	"fmt"
 
+	mgmtkeyvault "github.com/Azure/azure-sdk-for-go/services/keyvault/mgmt/2019-09-01/keyvault"
 	mgmtstorage "github.com/Azure/azure-sdk-for-go/services/storage/mgmt/2021-09-01/storage"
 	"github.com/Azure/go-autorest/autorest/to"
 
@@ -17,6 +18,9 @@ import (
 var (
 	storageAccountName         string = "parameters('oidcStorageAccountName')"
 	resourceTypeStorageAccount string = "Microsoft.Storage/storageAccounts"
+
+	SharedMSIKeyVaultName       = "concat(take(resourceGroup().name,10), '" + SharedMSIKeyVaultNameSuffix + "')"
+	SharedMSIKeyVaultNameSuffix = "-dev-msi"
 )
 
 func (g *generator) oicStorageAccount() *arm.Resource {
@@ -54,5 +58,22 @@ func (g *generator) oicRoleAssignment() *arm.Resource {
 		resourceTypeStorageAccount,
 		storageAccountName,
 		fmt.Sprintf("concat(%s, '/Microsoft.Authorization/', guid(resourceId('%s', %s)))", storageAccountName, resourceTypeStorageAccount, storageAccountName),
+	)
+}
+
+// devMSIKeyvault returns an arm.Resource representing a shared key vault to be used for persisting mock MSI certificates when
+// the RP is running in local development mode.
+func (g *generator) devMSIKeyvault() *arm.Resource {
+	return g.keyVault(fmt.Sprintf("[%s]", SharedMSIKeyVaultName), &[]mgmtkeyvault.AccessPolicyEntry{}, nil, nil)
+}
+
+// devMSIKeyvaultRBAC returns an arm.Resource representing a role assignment that grants the local development mode's mock RP identity
+// the KeyVaultSecretsOfficer role on the shared dev MSI key vault.
+func (g *generator) devMSIKeyvaultRBAC() *arm.Resource {
+	return rbac.ResourceRoleAssignment(
+		rbac.RoleKeyVaultSecretsOfficer,
+		"parameters('rpServicePrincipalId')",
+		"Microsoft.KeyVault/vaults",
+		SharedMSIKeyVaultName,
 	)
 }
