+++ pkg/deploy/generator/templates_miwi_dev.go
@@ -7,16 +7,23 @@ import (
 	"github.com/Azure/ARO-RP/pkg/util/arm"
 )
 
-func (g *generator) devMSIKeyvaultTemplate() *arm.Template {
+func (g *generator) miwiDevSharedTemplate() *arm.Template {
 	t := templateStanza()
 
 	t.Resources = append(t.Resources,
+		g.oicStorageAccount(),
+		g.oicRoleAssignment(),
 		g.devMSIKeyvault(),
 		g.devMSIKeyvaultRBAC(),
 	)
 
-	t.Parameters["rpServicePrincipalId"] = &arm.TemplateParameter{
-		Type: "string",
+	t.Parameters = map[string]*arm.TemplateParameter{
+		"rpServicePrincipalId": {
+			Type: "string",
+		},
+		"oidcStorageAccountName": {
+			Type: "string",
+		},
 	}
 
 	return t
